Model changes

+++ src/main/java/com/ti/dao/CustomerRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import com.ti.model.TCustomer;
 
 
-public interface CustomerRepository extends JpaRepository<TCustomer, Integer>, JpaSpecificationExecutor<TCustomer> {
+public interface CustomerRepository extends JpaRepository<TCustomer, Long>, JpaSpecificationExecutor<TCustomer> {
 
 }
